backend/requirements.txt
flask
firebase-admin
PyPDF2
pdfplumber
python-docx
google-cloud
yake
flask-cors